+++ backend/backend/playlists/routes/playlists.py
@@ -5,6 +5,7 @@ from ninja import Router, UploadedFile
 
 from backend.musics.models import Music
 from backend.utils.schemas import ErrorSchema
+from config.api.auth import InvalidTokenException, token_is_valid
 
 from ..models import Playlist
 from ..schemas import PlaylistSchemaIn, PlaylistSchemaOut, PlaylistSchemaUpdateIn
@@ -35,14 +36,16 @@ def get_playlist(request, id: uuid.UUID):
         return 500, {"status": 500, "message": "Internal server error", "full_message": str(e)}
 
 
-@router.post("/", response={201: PlaylistSchemaOut, 400: ErrorSchema})
+@router.post("/", response={201: PlaylistSchemaOut, 400: ErrorSchema, 401: ErrorSchema})
 def create_playlist(request, playlist: PlaylistSchemaIn, image: UploadedFile = None):
     try:
+        is_authenticated = token_is_valid(request, True)
+        if not is_authenticated.is_valid:
+            raise InvalidTokenException(f"You are not logged in!\n{is_authenticated.message}")
         payload_dict = playlist.dict()
         musics = payload_dict.pop("musics", [])
-        owner_is_artist = payload_dict.pop("owner_is_artist")
-        payload_dict["object_id"] = payload_dict.pop("owner_id")
-        owner_type = "artist" if owner_is_artist else "user"
+        payload_dict["object_id"] = is_authenticated.user.id
+        owner_type = "artist" if is_authenticated.user.is_artist() else "user"
         payload_dict["content_type"] = ContentType.objects.get(app_label=f"{owner_type}s", model=owner_type)
         playlist = Playlist.objects.create(
             **payload_dict,
@@ -51,17 +54,28 @@ def create_playlist(request, playlist: PlaylistSchemaIn, image: UploadedFile = N
         for music in musics:
             playlist.musics.add(music)
         return 201, playlist
+    except InvalidTokenException as e:
+        return 401, {"status": 401, "message": "Unauthorized", "full_message": str(e)}
     except Exception as e:
         return 400, {"status": 400, "message": "Bad request", "full_message": str(e)}
 
 
-@router.patch("/{id}", response={200: PlaylistSchemaOut, 400: ErrorSchema})
+@router.patch("/{id}", response={200: PlaylistSchemaOut, 400: ErrorSchema, 401: ErrorSchema, 404: ErrorSchema})
 def update_playlist(request, id: uuid.UUID, playlist: PlaylistSchemaUpdateIn = None, image: UploadedFile = None):
     try:
+        is_authenticated = token_is_valid(request, True)
+        if not is_authenticated.is_valid:
+            raise InvalidTokenException(f"You are not logged in!\n{is_authenticated.message}")
         playlist_dict = playlist.dict(exclude_unset=True)
         if playlist_dict == {} and image is None:
             raise Exception("Data not provided")
         playlist = Playlist.objects.get(id=id)
+        if playlist.object_id != is_authenticated.user.id:
+            return 403, {
+                "status": 403,
+                "message": "Forbidden",
+                "full_message": "You are not the owner of this playlist",
+            }
         for key, value in playlist_dict.items():
             setattr(playlist, key, value)
         if image:
@@ -70,26 +84,51 @@ def update_playlist(request, id: uuid.UUID, playlist: PlaylistSchemaUpdateIn = N
         return 200, playlist
     except Playlist.DoesNotExist:
         return 404, {"status": 404, "message": "Playlist not found", "full_message": "Playlist not found"}
+    except InvalidTokenException as e:
+        return 401, {"status": 401, "message": "Unauthorized", "full_message": str(e)}
     except Exception as e:
         return 400, {"status": 400, "message": "Bad request", "full_message": str(e)}
 
 
-@router.delete("/{id}", response={204: None, 404: ErrorSchema, 500: ErrorSchema})
+@router.delete("/{id}", response={204: None, 401: ErrorSchema, 403: ErrorSchema, 404: ErrorSchema, 500: ErrorSchema})
 def delete_playlist(request, id: uuid.UUID):
     try:
+        is_authenticated = token_is_valid(request, True)
+        if not is_authenticated.is_valid:
+            raise InvalidTokenException(f"You are not logged in!\n{is_authenticated.message}")
         playlist = Playlist.objects.get(id=id)
+        if playlist.object_id != is_authenticated.user.id:
+            return 403, {
+                "status": 403,
+                "message": "Forbidden",
+                "full_message": "You are not the owner of this playlist",
+            }
         playlist.delete()
         return 204, None
     except Playlist.DoesNotExist:
         return 404, {"status": 404, "message": "Playlist not found", "full_message": "Playlist not found"}
+    except InvalidTokenException as e:
+        return 401, {"status": 401, "message": "Unauthorized", "full_message": str(e)}
     except Exception as e:
         return 500, {"status": 500, "message": "Internal server error", "full_message": str(e)}
 
 
-@router.patch("/{id}/add_musics", response={200: PlaylistSchemaOut, 400: ErrorSchema, 404: ErrorSchema})
+@router.patch(
+    "/{id}/add_musics",
+    response={200: PlaylistSchemaOut, 400: ErrorSchema, 401: ErrorSchema, 403: ErrorSchema, 404: ErrorSchema},
+)
 def add_musics_to_playlist(request, id: uuid.UUID, musics: list[uuid.UUID]):
     try:
+        is_authenticated = token_is_valid(request, True)
+        if not is_authenticated.is_valid:
+            raise InvalidTokenException(f"You are not logged in!\n{is_authenticated.message}")
         playlist = Playlist.objects.get(id=id)
+        if playlist.object_id != is_authenticated.user.id:
+            return 403, {
+                "status": 403,
+                "message": "Forbidden",
+                "full_message": "You are not the owner of this playlist",
+            }
         for music in musics:
             music_obj = Music.objects.filter(id=music).first()
             if not music_obj:
@@ -101,19 +140,35 @@ def add_musics_to_playlist(request, id: uuid.UUID, musics: list[uuid.UUID]):
         return 200, playlist
     except Playlist.DoesNotExist:
         return 404, {"status": 404, "message": "Playlist not found", "full_message": "Playlist not found"}
+    except InvalidTokenException as e:
+        return 401, {"status": 401, "message": "Unauthorized", "full_message": str(e)}
     except Exception as e:
         return 400, {"status": 400, "message": "Bad request", "full_message": str(e)}
 
 
-@router.patch("/{id}/remove_musics", response={200: PlaylistSchemaOut, 400: ErrorSchema, 404: ErrorSchema})
+@router.patch(
+    "/{id}/remove_musics",
+    response={200: PlaylistSchemaOut, 400: ErrorSchema, 401: ErrorSchema, 403: ErrorSchema, 404: ErrorSchema},
+)
 def remove_musics_from_playlist(request, id: uuid.UUID, musics: list[uuid.UUID]):
     try:
+        is_authenticated = token_is_valid(request, True)
+        if not is_authenticated.is_valid:
+            raise InvalidTokenException(f"You are not logged in!\n{is_authenticated.message}")
         playlist = Playlist.objects.get(id=id)
+        if playlist.object_id != is_authenticated.user.id:
+            return 403, {
+                "status": 403,
+                "message": "Forbidden",
+                "full_message": "You are not the owner of this playlist",
+            }
         for music in musics:
             playlist.musics.remove(music)
         playlist.save()
         return 200, playlist
     except Playlist.DoesNotExist:
         return 404, {"status": 404, "message": "Playlist not found", "full_message": "Playlist not found"}
+    except InvalidTokenException as e:
+        return 401, {"status": 401, "message": "Unauthorized", "full_message": str(e)}
     except Exception as e:
         return 400, {"status": 400, "message": "Bad request", "full_message": str(e)}
